requirements.txt
chromadb==0.5
fastapi==0.104.1
gitpython==3.1.40
httpx==0.25.2
huggingface-hub==0.25.2
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-ollama
langgraph
numpy==1.26.4
prometheus-client==0.19.0
pydantic==2.7.4
pygithub==2.1.1
pytest-asyncio==0.21.1
pytest==7.4.3
python-dotenv==1.0.0
redis==5.0.1
sentence-transformers==2.2.2
uvicorn==0.24.0